+++ src/main/java/md/varoinform/model/search/CreationDateSearcher.java
@@ -15,7 +15,7 @@ import java.util.regex.Pattern;
  */
 public class CreationDateSearcher extends Searcher {
     @Override
-    public List<Long> search(String q) {
+    public List<Integer> search(String q) {
         try (DefaultClosableSession session = new DefaultClosableSession()) {
             String hql = "Select distinct e.id from Enterprise e where e.creation = :creation";
             if (!Pattern.matches("^[0-9]+$", q.trim())) return new ArrayList<>();
