+++ src/main/java/md/varoinform/model/search/EmailsSearcher.java
@@ -13,7 +13,7 @@ import java.util.List;
  */
 public class EmailsSearcher extends Searcher {
     @Override
-    public List<Long> search(String q) {
+    public List<Integer> search(String q) {
         try (DefaultClosableSession session = new DefaultClosableSession()) {
             String hql = "Select distinct e.id from Enterprise e join e.contacts cs join cs.emails email where lower(email.email) = :pattern";
             Query query = session.createQuery(hql).setString("pattern", q.trim().toLowerCase());
