+++ src/main/java/md/varoinform/model/search/SectorSearcher.java
@@ -14,7 +14,7 @@ import java.util.List;
  */
 public class SectorSearcher extends Searcher {
     @Override
-    public List<Long> search(String q) {
+    public List<Integer> search(String q) {
         try (DefaultClosableSession session = new DefaultClosableSession()) {
             String field = "titles.title";
             Normalizer normalizer = new Normalizer(field, q, Normalizer.RO);
