+++ src/main/java/md/varoinform/view/demonstrator/EnterpriseTransferableHandler.java
@@ -25,7 +25,7 @@ public class EnterpriseTransferableHandler extends TransferHandler {
     @Override
     protected Transferable createTransferable(JComponent component) {
         TableView tableView = (TableView) component;
-        List<Long> selected = tableView.getSelected();
+        List<Integer> selected = tableView.getSelected();
         if (selected.size() <= 0) return null;
         return new EnterpriseTransferable(selected);
     }
@@ -43,7 +43,7 @@ public class EnterpriseTransferableHandler extends TransferHandler {
         }
         try {
             @SuppressWarnings("unchecked")
-            List<Long> selected = (List<Long>) t.getTransferData(EnterpriseTransferable.ENTERPRISE_FLAVOR);
+            List<Integer> selected = (List<Integer>) t.getTransferData(EnterpriseTransferable.ENTERPRISE_FLAVOR);
             JList.DropLocation dropLocation = list.getDropLocation();
             Point point = dropLocation.getDropPoint();
             int index = list.locationToIndex(point);
